requirements.txt
pyrogram
tgcrypto
motor
python-dotenv
uvloop
